requirements.txt
squacapipy
psycopg2-binary
flake8
requests